Rename editSurvey and createSurvey
