+++ client/src/containers/EditSurveyPanel.js
@@ -1,13 +1,13 @@
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import EditSurveyPanelView from '../components/EditPanel/EditSurveyPanel'
-import { getSurvey } from '../reducers/edit_survey'
-import { updateSurveyHeader } from '../actions/edit_survey'
+import { getSurvey } from '../reducers/editSurvey'
+import { updateSurveyHeader } from '../actions/editSurvey'
 
 const mapStateToProps = (state) => {
   return {
-    title: getSurvey(state.edit_survey).title,
-    subTitle: getSurvey(state.edit_survey).subTitle
+    title: getSurvey(state.editSurvey).title,
+    subTitle: getSurvey(state.editSurvey).subTitle
   }
 }
