+++ client/src/pages/EditSurveyPage.js
@@ -6,7 +6,7 @@ import SurveyPreview from './../containers/SurveyPreview/SurveyPreview'
 import EditTab from './../containers/EditTab'
 import EditFooter from './../containers/EditFooter'
 
-import { fetchSurvey } from '../actions/edit_survey'
+import { fetchSurvey } from '../actions/editSurvey'
 
 import { AutoAffix } from 'react-overlays'
