+++ client/src/pages/UserSurveyPage.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import SurveyNavBar from './../containers/SurveyNavBar'
-import { fetchSurvey } from '../actions/edit_survey'
+import { fetchSurvey } from '../actions/editSurvey'
 import './EditSurveyPage.css'
 
 class EditSurveyPage extends React.Component {
