+++ client/src/reducers/index.js
@@ -3,8 +3,8 @@ import { combineReducers } from 'redux'
 import register from './register'
 import session from './session'
 import surveys from './surveys'
-import createSurvey from './create_survey'
-import editSurvey from './edit_survey'
+import createSurvey from './createSurvey'
+import editSurvey from './editSurvey'
 import survey from './survey'
 import data from './data/index'
